Move error file to the public utility folder
