+++ include/cuco/detail/static_map.inl
@@ -1,5 +1,5 @@
 /*
- * Copyright (c) 2020-2022, NVIDIA CORPORATION.
+ * Copyright (c) 2020-2023, NVIDIA CORPORATION.
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -15,9 +15,9 @@
  */
 
 #include <cuco/detail/bitwise_compare.cuh>
-#include <cuco/detail/error.hpp>
 #include <cuco/detail/utils.cuh>
 #include <cuco/detail/utils.hpp>
+#include <cuco/utility/error.hpp>
 
 #include <thrust/iterator/transform_iterator.h>
 #include <thrust/iterator/zip_iterator.h>
