+++ include/cuco/dynamic_map.cuh
@@ -17,10 +17,10 @@
 #pragma once
 
 #include <cuco/detail/dynamic_map_kernels.cuh>
-#include <cuco/detail/error.hpp>
 #include <cuco/hash_functions.cuh>
 #include <cuco/sentinel.cuh>
 #include <cuco/static_map.cuh>
+#include <cuco/utility/error.hpp>
 
 #include <thrust/device_vector.h>
 #include <thrust/functional.h>
